Remove the Directory type.
I'd removed the recursive functionality of the Directory type, and it
was becoming just a placeholder for what should've been
module-level functions.

I'm going to remove it for now, and just provide module-level
functions instead.

When we bring back recursive/directory-based functionality,
(ex: walkdir) we can add a Directory type back.