Fix types for TypeScript 5.7 (Deno 2.2)
Deno v2.2 upgraded to TypeScript 5.7.

There's a nice writeup of the relevant change in this PR:
https://github.com/microsoft/TypeScript/pull/59417

In short, there are now two variants of Uint8Array:
Uint8Array<ArrayBuffer> and Uint8Array<SharedArrayBuffer>.

And their types have diverged. If you don't specify which one you're
returning, you get a union of the two, which is no longer
assignment-compatible with ArrayBuffer. 

Fix our types (and some Deno types) to be explicit there.